03. code/requirements.txt
langchain==0.3.27
langchain-openai==0.3.32
langchain-community==0.3.29
langchain-aws==0.2.31
langchain-anthropic==0.3.19
python-dotenv==1.1.1
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
boto3==1.34.0
pydantic==2.5.0
